Add code for squaring off genotypes that include reference models.
Resolves bigdatagenomics/adam#1312. Enables gVCF style data to be used with the
joint variant caller by extracting the sites where a variant was called as a
genotype in one of the gVCFs. These variant sites are then joined back against
the original genotypes. If the genotyped allele was not present and a reference
model block is present, the allele is extracted out from the reference model.